Issue 142 fixed. NumRaters was only parsing an int16, not a larger number

+++ trunk/clients/cs/src/extensions/ratings.cs
@@ -89,7 +89,7 @@ namespace Google.GData.Extensions
         {
             get
             {
-                return Int16.Parse(this.Attributes[GDataParserNameTable.XmlAttributeNumRaters] as string);
+                return Int32.Parse(this.Attributes[GDataParserNameTable.XmlAttributeNumRaters] as string);
             }
             set
             {
